requirements.txt
# \python Core Requirements
# Only packages actually used by \python self-authored code

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
uvicorn>=0.27.0
aiohttp>=3.9.0
starlette>=0.36.0
fastmcp>=0.1.0

# HTTP
requests>=2.31.0
httpx>=0.27.0

# Data
pyyaml>=6.0
pillow>=10.0.0
sqlalchemy>=2.0.0

# Crypto & Auth
cryptography>=42.0.0

# AI/ML
openai>=1.12.0
huggingface_hub>=0.21.0

# MCP
mcp>=1.0.0

# Security Tools
bandit>=1.7.9
detect-secrets>=1.5.0
pip-audit>=0.7.0
safety>=3.2.0

# Security Libraries (replacing unsafe patterns)
defusedxml>=0.7.1
simpleeval>=1.0.0
keyring>=25.0.0
certifi>=2024.0.0

# Utils
python-dotenv>=1.0.0
packaging>=24.0
